feat: add PUT favorites & GET favorites
Added the gateway and service to PUT a crous to favorites, also added a query for the GET of findall to decide if the API returns only the favorites or everything. The service toggleFavorite will add a crous to favorites if not added before, and remove it if added before. Added test for the PUT also.